Restructuring the namespaces for Interfaces and Services
