+++ src/Core/Interfaces/Repositories/ICaretakerRepository.cs
@@ -1,6 +1,6 @@
 using Neurocorp.Api.Core.Entities;
 
-namespace Neurocorp.Api.Core.Interfaces;
+namespace Neurocorp.Api.Core.Interfaces.Repositories;
 
 public interface ICaretakerRepository : IRepository<Caretaker>
 {
