+++ src/Core/Interfaces/Services/IPatientService.cs
@@ -2,9 +2,9 @@ using Neurocorp.Api.Core.Entities;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace Neurocorp.Api.Core.Interfaces;
+namespace Neurocorp.Api.Core.Interfaces.Services;
 
-public interface ITherapistService : IService<Therapist>
+public interface IPatientService : IService<Patient>
 {
     // other business logic operations...
 }
