+++ src/Core/Interfaces/Services/ITherapistService.cs
@@ -2,9 +2,9 @@ using Neurocorp.Api.Core.Entities;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace Neurocorp.Api.Core.Interfaces;
+namespace Neurocorp.Api.Core.Interfaces.Services;
 
-public interface IPatientService : IService<Patient>
+public interface ITherapistService : IService<Therapist>
 {
     // other business logic operations...
 }
