+++ src/Core/Interfaces/Services/ITherapySessionService.cs
@@ -2,7 +2,7 @@ using Neurocorp.Api.Core.Entities;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace Neurocorp.Api.Core.Interfaces;
+namespace Neurocorp.Api.Core.Interfaces.Services;
 
 public interface ITherapySessionService : IService<TherapySession>
 {
