+++ src/Core/Services/PatientProfileService.cs
@@ -1,6 +1,7 @@
-using Neurocorp.Api.Core.BusinessObjects;
+using Neurocorp.Api.Core.BusinessObjects.Patients;
 using Neurocorp.Api.Core.Entities;
-using Neurocorp.Api.Core.Interfaces;
+using Neurocorp.Api.Core.Interfaces.Services;
+using Neurocorp.Api.Core.Interfaces.Repositories;
 using System.Collections.Generic;
 using System.Threading.Tasks;
