+++ src/Core/Services/TherapistService.cs
@@ -1,5 +1,6 @@
 using Neurocorp.Api.Core.Entities;
-using Neurocorp.Api.Core.Interfaces;
+using Neurocorp.Api.Core.Interfaces.Repositories;
+using Neurocorp.Api.Core.Interfaces.Services;
 using System.Collections.Generic;
 using System.Threading.Tasks;
