+++ src/Infrastructure/Repositories/PatientProfileRepository.cs
@@ -1,5 +1,6 @@
 using Neurocorp.Api.Core.Entities;
-using Neurocorp.Api.Core.Interfaces;
+using Neurocorp.Api.Core.BusinessObjects.Patients;
+using Neurocorp.Api.Core.Interfaces.Repositories;
 using Neurocorp.Api.Infrastructure.Data;
 using Microsoft.EntityFrameworkCore;
 using Neurocorp.Api.Core.BusinessObjects;
