+++ src/Infrastructure/Repositories/UserRepository.cs
@@ -1,5 +1,5 @@
 using Neurocorp.Api.Core.Entities;
-using Neurocorp.Api.Core.Interfaces;
+using Neurocorp.Api.Core.Interfaces.Repositories;
 using Neurocorp.Api.Infrastructure.Data;
 
 namespace Neurocorp.Api.Infrastructure.Repositories;
