+++ tests/Web.Tests/PatientsControllerTests.cs
@@ -1,8 +1,8 @@
 using Moq;
 using Xunit;
 using Microsoft.AspNetCore.Mvc;
-using Neurocorp.Api.Core.BusinessObjects;
-using Neurocorp.Api.Core.Interfaces;
+using Neurocorp.Api.Core.BusinessObjects.Patients;
+using Neurocorp.Api.Core.Interfaces.Services;
 using Neurocorp.Api.Web.Controllers;
 using System.Collections.Generic;
 using System.Threading.Tasks;
